test(analysis): fix expected results
The expected results count on the fact that the DB lookup phase and
the in-memory phase are aligned. Which they currently are not.

For default (search everything) the in-memory part take the PURL into
account, while the DB phase does not. This is an issue with the `q`
implementation and should be fixed in another PR. However, even after
the fix, this would result in the same number of returned items.

The reason whe there hard been more, is because the graph loading logic
did load more related SBOMs than required. Which then ended up in the
pool of SBOMs to be searched.